order-service/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
nacos-sdk-python==0.1.5
requests==2.31.0
mysqlclient==2.2.0
python-dotenv==1.0.0
cryptography==41.0.7
